+++ wintasee/wintasee.cpp
@@ -608,7 +608,7 @@ void MakeWindowWindowed(HWND hwnd, DWORD width, DWORD height)
 {
 //	if(hwndSizeLocked.find(hwnd) == hwndSizeLocked.end())
 //		return;
-	RECT rect = {0,0,width,height};
+	RECT rect = {0, 0, static_cast<LONG>(width), static_cast<LONG>(height)};
 	AdjustWindowRect(&rect, WS_OVERLAPPEDWINDOW, FALSE);
 	SetWindowLong(hwnd, GWL_STYLE, WS_OVERLAPPEDWINDOW);
 	//SetWindowPos(hwnd, 0, rect.left, rect.top, rect.right - rect.left, rect.bottom - rect.top, /*SWP_NOMOVE | */SWP_SHOWWINDOW);
